--- Demo/CMakeLists.txt
@@ -1,6 +1,0 @@
-# Common demo library
-add_subdirectory(Lib)
-
-# Demos
-add_subdirectory(Ballistic)
-add_subdirectory(BigBallistic)
