+++ Demos/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Common demo library
+add_subdirectory(Lib)
+
+# Demos
+add_subdirectory(Ballistic)
+add_subdirectory(BigBallistic)
+add_subdirectory(Blob)
+add_subdirectory(Bridge)
+add_subdirectory(Explosion)
