+++ Demos/Lib/CMakeLists.txt
@@ -33,3 +33,7 @@ target_compile_options(${PROJECT_NAME} PRIVATE
   $<$<NOT:$<CXX_COMPILER_ID:MSVC>>:-Wall -Werror>
   # -Wextra -Wpedantic 
 )
+
+set_target_properties(${PROJECT_NAME} 
+	PROPERTIES FOLDER "Demos"
+)
